indexer: rename testing
