+++ packages/indexer/src/testing/index.ts
@@ -1,1 +1,2 @@
 export * from "./setup";
+export * from "./vcr";
